+++ app/src/main/java/com/vlad/bikegarage/rides/presentation/addride/AddRideViewModel.kt
@@ -59,7 +59,7 @@ class AddRideViewModel @AssistedInject constructor(
             is AddRideEvent.OnRideDateAdded -> {}
             is AddRideEvent.OnRideDistanceAdded -> {
                 _state.update { newState ->
-                    newState.copy(distance = event.distance)
+                    newState.copy(distance = event.distance.toInt())
                 }
             }
